POC of using Context class instead Activity class to init ZeroPush
